Simplify AoC panel item creation, add AoC 2024 solution code file folder to assets to fix UI error, fix AoC 2024 day 1 part 2 result being a double instead of int

+++ lib/solvers/advent_of_code/2024/day_01_solver.dart
@@ -19,7 +19,7 @@ class Day01Solver extends AdventOfCode2024Solver {
     List<int> list1 = locationIdGroups.map((group) => group[0]).toList();
     List<int> list2 = locationIdGroups.map((group) => group[1]).toList();
 
-    double totalSimilarity = list1.foldIndexed(0, (index, previous, element) => previous + element * list2.countValueOccurances(element));
+    int totalSimilarity = list1.foldIndexed(0, (index, previous, element) => previous + element * list2.countValueOccurances(element));
 
     return 'Total distance: $totalDistance, total similarity: $totalSimilarity';
   }
